Add another assertion for aria-modal=true
ARIA 1.1 does not require that AT obscure content outside of the modal. That is only a 'may' in the spec, and authors should not depend on it. A new assertion was added that the true value be conveyed, either via an explicit announcement or implied by other content being obscured.