make the default pass strategy 'all'
'any' led to fairly misleading aggregate data in some cases. In practice the two result in similar aggregate results in most cases. However, it is difficult to defend fully 'passing' an assertion when a role, name, value, etc is not announced for one command but is announced for another command.